[fe] feat (#66) geoLocation 함수 분리 - geoLocation을 사용할 일이 많아 utils로 분리
