+++ client/src/utils/getLocation.ts
@@ -1,0 +1,28 @@
+import { Location } from '@custom-types';
+
+const getLocation = (): Promise<Location> => {
+  return new Promise<Location>((resolve, reject) => {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition(
+        (position) => {
+          resolve({
+            lat: position.coords.latitude,
+            lng: position.coords.longitude,
+          });
+        },
+        (error) => {
+          reject(error);
+        },
+        {
+          enableHighAccuracy: false,
+          maximumAge: 0,
+          timeout: Infinity,
+        },
+      );
+    } else {
+      reject(new Error('Unable to retrieve your location'));
+    }
+  });
+};
+
+export default getLocation;
